+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added executables whitelist
+
 ## v0.4.20 (2024-08-09)
 * Ignore node_modules in executable permissions check